+++ src/test/java/org/folio/rest/impl/WireMockTestBase.java
@@ -1,5 +1,12 @@
 package org.folio.rest.impl;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.badRequest;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.ok;
+import static com.github.tomakehurst.wiremock.client.WireMock.serverError;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static io.restassured.RestAssured.given;
 import static org.apache.http.HttpStatus.SC_BAD_REQUEST;
 import static org.apache.http.HttpStatus.SC_CREATED;
@@ -7,9 +14,12 @@ import static org.apache.http.HttpStatus.SC_NO_CONTENT;
 import static org.apache.http.HttpStatus.SC_OK;
 import static org.folio.repository.kbcredentials.KbCredentialsTableConstants.KB_CREDENTIALS_TABLE_NAME;
 import static org.folio.rest.util.RestConstants.JSON_API_TYPE;
+import static org.folio.service.locale.LocaleSettingsServiceImpl.LOCALE_ENDPOINT_PATH;
+import static org.folio.test.util.TestUtil.readFile;
 import static org.folio.util.KbTestUtil.clearDataFromTable;
 import static org.hamcrest.Matchers.notNullValue;
 
+import com.github.tomakehurst.wiremock.http.Fault;
 import io.restassured.RestAssured;
 import io.restassured.http.Header;
 import io.restassured.http.Headers;
@@ -19,6 +29,7 @@ import io.vertx.ext.unit.TestContext;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import lombok.SneakyThrows;
 import org.apache.http.HttpStatus;
 import org.apache.http.protocol.HTTP;
 import org.folio.cache.VertxCache;
@@ -199,6 +210,42 @@ public abstract class WireMockTestBase extends TestBase {
       .then()
       .statusCode(SC_BAD_REQUEST)
       .body("errors.first.title", notNullValue());
+  }
+
+  protected void mockSuccessfulLocaleResponse() {
+    mockSuccessfulLocaleResponse("responses/configuration/locale-settings.json");
+  }
+
+  @SneakyThrows
+  protected void mockSuccessfulLocaleResponse(String configFileName) {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(ok(readFile(configFileName))));
+  }
+
+  @SneakyThrows
+  protected void mockFailedLocaleResponse() {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(badRequest()));
+  }
+
+  protected void mockLocaleResponseWithInvalidJson() {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(ok("{ invalid json }")));
+  }
+
+  protected void mockLocaleResponseWithEmptyBody() {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(ok()));
+  }
+
+  protected void mockLocaleResponseWithServerError() {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(serverError()));
+  }
+
+  protected void mockLocaleResponseWithNetworkError() {
+    stubFor(get(urlPathEqualTo(LOCALE_ENDPOINT_PATH))
+      .willReturn(aResponse().withFault(Fault.CONNECTION_RESET_BY_PEER)));
   }
 
   private Header[] addContentHeader(Header[] headers) {